+++ e2e/config-esm-mjs/cypress/support/e2e.mjs
@@ -1,0 +1,9 @@
+import registerCypressGrep from "@cypress/grep/src/support";
+require("cypress-terminal-report/src/installLogsCollector")();
+require("cypress-cloud/support");
+require("./commands");
+
+registerCypressGrep();
+beforeEach(() => {
+  cy.visit("/");
+});