+++ packages/cypress-cloud/lib/config/config.ts
@@ -1,10 +1,10 @@
 import Debug from "debug";
 import path from "path";
+import os from 'os'
+import { P, match } from "ts-pattern";
 import { DetectedBrowser, ValidatedCurrentsParameters } from "../../types";
 import { bootCypress } from "../bootstrap";
 import { warn } from "../log";
-import { require } from "../require";
-import { getRandomPort } from "../utils";
 
 const debug = Debug("currents:config");
 
@@ -24,36 +24,56 @@ export type CurrentsConfig = {
 
 let _config: CurrentsConfig | null = null;
 
-export function getCurrentsConfig(projectRoot?: string): CurrentsConfig {
+const defaultConfig: CurrentsConfig = {
+  e2e: {
+    batchSize: 3,
+  },
+  component: {
+    batchSize: 5,
+  },
+  cloudServiceUrl: "https://cy.currents.dev",
+};
+
+export async function getCurrentsConfig(
+  projectRoot?: string
+): Promise<CurrentsConfig> {
   if (_config) {
     return _config;
   }
-  const defaultConfig: CurrentsConfig = {
-    e2e: {
-      batchSize: 3,
-    },
-    component: {
-      batchSize: 5,
-    },
-    cloudServiceUrl: "https://cy.currents.dev",
-  };
 
   const configFilePath = getConfigFilePath(projectRoot);
-  try {
-    const resolvedPath = path.resolve(...configFilePath);
-    debug("loading currents config file from '%s'", resolvedPath);
+  // try loading possible config files
+  for (const filepath of configFilePath) {
+    const config = match(await loadConfigFile(filepath))
+      .with({ default: P.not(P.nullish) }, (c) => c.default)
+      .with(P.not(P.nullish), (c) => c)
+      .otherwise(() => null);
 
-    const fsConfig = require(resolvedPath);
-    _config = {
-      ...defaultConfig,
-      ...fsConfig,
-    } as CurrentsConfig;
-    return _config;
+    if (config) {
+      debug("loaded currents config from '%s'\n%O", filepath, config);
+      _config = {
+        ...defaultConfig,
+        ...config,
+      };
+      return _config;
+    }
+  }
+
+  warn(
+    "Failed to load config file, falling back to the default config. Attempted locations: %s",
+    configFilePath
+  );
+  _config = defaultConfig;
+  return _config;
+}
+
+async function loadConfigFile(filepath: string) {
+  try {
+    debug("loading currents config file from '%s'", filepath);
+    return await import(filepath);
   } catch (e) {
-    warn("failed to load config file: %s", configFilePath);
-    debug("failure details: %s", e);
-    _config = defaultConfig;
-    return _config;
+    debug("failed loading config file from: %s", e);
+    return null;
   }
 }
 
@@ -66,7 +86,7 @@ export async function getMergedConfig(params: ValidatedCurrentsParameters) {
         rawJson: Record<string, unknown>;
         browsers: DetectedBrowser[];
       })
-    | undefined = await bootCypress(getRandomPort(), params);
+    | undefined = await bootCypress(params);
 
   debug("cypress resolvedConfig: %O", cypressResolvedConfig);
 
@@ -94,6 +114,11 @@ export async function getMergedConfig(params: ValidatedCurrentsParameters) {
   return result;
 }
 
-function getConfigFilePath(projectRoot: string | null = null) {
-  return [projectRoot ?? process.cwd(), "currents.config.js"];
+function getConfigFilePath(projectRoot: string | null = null): string[] {
+  const prefix = projectRoot ?? process.cwd();
+  return [
+    "currents.config.js",
+    "currents.config.cjs",
+    "currents.config.mjs",
+  ].map((p) => `file://${path.resolve(prefix, p)}`);
 }