+++ packages/cypress-cloud/lib/results/uploadResults.ts
@@ -7,41 +7,13 @@ import {
   UpdateInstanceResultsPayload,
 } from "../api";
 import { uploadArtifacts, uploadStdoutSafe } from "../artifacts";
+import { setCancellationReason } from "../cancellation";
 import { getInitialOutput } from "../capture";
 import { isCurrents } from "../env";
-import { warn } from "../log";
-import { setCancellationReason } from "../state";
 import { getInstanceResultPayload, getInstanceTestsPayload } from "./results";
 const debug = Debug("currents:results");
 
-export async function getUploadResultsTask({
-  instanceId,
-  spec,
-  runResult,
-  output,
-}: {
-  instanceId: string;
-  spec: string;
-  runResult: CypressCommandLine.CypressRunResult;
-  output: string;
-}) {
-  const run = runResult.runs.find((r) => r.spec.relative === spec);
-  if (!run) {
-    warn('Cannot determine run result for spec "%s"', spec);
-    return;
-  }
-  return processCypressResults(
-    instanceId,
-    {
-      // replace the runs with the run for the specified spec
-      ...runResult,
-      runs: [run],
-    },
-    output
-  );
-}
-
-export async function processCypressResults(
+export async function getReportResultsTask(
   instanceId: string,
   results: CypressCommandLine.CypressRunResult,
   stdout: string
@@ -63,6 +35,7 @@ export async function processCypressResults(
     debug("instance %s should cancel", instanceId);
     setCancellationReason(cloud.shouldCancel);
   }
+
   debug("instance %s artifact upload instructions %o", instanceId, {
     videoUploadUrl,
     screenshotUploadUrls,